SQLtoSheetsTool/requirements.txt
pymysql
google-auth
google-auth-oauthlib
google-api-python-client
pandas